Some modifications

+++ src/config.rs
@@ -8,7 +8,7 @@ use std::io::Read;
 pub struct Config {
     pub glitch_private_key: Option<String>,
     pub glitch_fee_address: String,
-    pub interval_days_for_transfer: u8,
+    pub interval_days_for_transfer: u32,
     pub business_fee: f64,
     pub glitch_gas: bool,
     pub db: Database,
